+++ README.md
@@ -1,6 +1,8 @@
-# Alpaka - A Danish Instruction-Following Dataset Generation Recipe
+# Lærebogen - A Danish Instruction-Following Dataset Generation Recipe
 
-This is a Danish version of the English Alpaca dataset generation recipe, which can be
-used to generate Danish instruction-following datasets using generative base models
-hosted on Ollama. We also release a Danish instruction-following dataset, Alpaka,
-containing X instructions and responses.
+This is a Danish instruction-following dataset generation recipe, which is based on the
+English Alpaca/SelfInstruct and EvolInstruct dataset generation recipes.
+
+The dataset is generated using commercially viable large language models (LLMs) via the
+Ollama inference framework, enabling a permissive dataset license, in contrast to the
+English Alpaca dataset.
