--- src/alpaka/__init__.py
@@ -1,1 +1,0 @@
-"""Alpaka: A Danish Instruction-Following Dataset Generation Recipe."""
