+++ src/laerebogen/__init__.py
@@ -1,0 +1,1 @@
+"""A Danish Instruction-Following Dataset Generation Recipe."""
